fix: commonize resas api requesting for source-code-optimization
